Run benchmark via `execa`
Previously, we saw issues where the results were ending up varying
wildly for some cases, likely because we were ending up with a dirty
heap, triggering GC, and so on. Now, run in a new process each time;
this produces much more consistent results. Curiously, it also shows
much higher run times, but the proportions are consistent with those
from previous analyses.

Combined with the changes in a previous commit to generate a table, and
adding in min and max values, the results currently look like this when
averaged over five separate runs for each source we care about:

|           variant            | average |  min  |  max  |
| ---------------------------- | ------- | ----- | ----- |
| ember-source@3.24.5          | 14381.6 | 13538 | 14790 |
| ember-source@3.25.4          | 44363.6 | 43375 | 45090 |
| ember-source@3.28.1          | 45368.2 | 45224 | 45484 |
| @glimmer/compiler@0.65.3     | 14493.8 | 14353 | 14600 |
| @glimmer/compiler@0.80.0     | 44395   | 44018 | 44679 |
| @glimmer/compiler@experiment | 20633.8 | 20523 | 20777 |
| @glimmer/compiler@latest     | 20751.4 | 20569 | 20864 |